Finishing some bugfixes; the code is fully functional again.

+++ MultiKeys/Remapper/Modifier.cpp
@@ -53,7 +53,7 @@ namespace Multikeys
 	bool CompositeModifier::matches(Scancode sc) const
 	{
 		// An incoming scancode only needs to match one of this modifier's scancodes.
-		for (unsigned short i = 0; i < scanCount; i++)
+		for (size_t i = 0; i < scanCount; i++)
 		{
 			if (scanArray[i] == sc)
 				return true;
@@ -107,27 +107,30 @@ namespace Multikeys
 
 	bool ModifierStateMap::checkState(Level* const level) const
 	{
-		// 1. Every name present in modNames must be set to true in this->modifiers
-		// 2. Every name not present in modNames must be set to false in this->modifiers
+		// 0. Level's modNames contains precisely the modifiers that trigger it.
+		// 1. Every name present in modNames must be already set to true in this->modifiers
+		// 2. Every name not present in modNames must be false in this->modifiers
 		// Satified those conditions, this method must return true
 		// Not satified those conditions, this method must return false.
 		size_t i = 0;
 		bool found = false;
 
-		// Loop through list of all modifiers
+		// Loop through all modifiers in this map
 		for (auto it = this->modifiers.begin(); it != this->modifiers.end(); it++)
 		{
 			// Determine whether or not this modifier is present in the Level's modifiers
 			// Look for it by name
 			for (i = 0, found = false; i < level->modifierCombination.size(); i++)
 			{
-				if (level->modifierCombination[i] == it->first->name)
+				if (level->modifierCombination[i] == it->first->name)	// comparing wstrings
 				{
 					found = true;
-					break;	// get out of for loop
+					break;	// get out of this inner loop
 				}
 			}
 			// Presence in the Level's modifers must be equal to the modifier's current state
+			// That is, if the modifier exists in the Level, then it must be currently pressed
+			// if the modifier was not found in the Level, then it must be currently unpressed.
 			if (found != it->second)
 			{
 				return false;
@@ -143,6 +146,7 @@ namespace Multikeys
 			it->second = false;
 	}
 
+	// This is currently not in use, I believe.
 	inline bool ModifierStateMap::operator==(const ModifierStateMap& rhs) const
 	{
 		return rhs.modifiers.size() == this->modifiers.size()
